Semi-significant changes to try to resolve user permission issues.
We're still connecting under our own personal user accounts to the
servers, so permissions can be granted based our normal group
membership, but now we immediately sudo to a single "deploy" user and
attempt to do everything as that single user. This will hopefully
resolve pesky chown/chgrp permission issues.

The other part of this is that we're no longer silently ignoring certain
chmod/chown/chgrp errors. Previously, we assumed this failed because
the permissions were already correct, but that was not always the case.
Now we fail if these permissions changing commands fail.

This also includes a few small cleanups and additions:

- On deployment lock errors, show when the lock file was created to help
  identify stale locks.
- Add timers and logging to the end to show how long the deployment
  took.