Bring back phased deployments for torquebox.
This is mostly based on the Torquebox 2 functionality we got rid of in
429ff7e43fb5d9a4cea8ef14f03fa973ded6319c. This brings back the
zero-downtime deployment approach where each deployment maintains its
own version of the app in torquebox, and then we explicitly undeploy the
older versions after deployment of the new version is finished.

This is in attempt to workaround memory leaks we were experiencing with
the Torquebox 3's built-in hot deploy approach. But we're still
experimenting, so this new code allows you to deploy in either way
("phased" for the multi-release style or "hot" for the single release
 touch tmp/restart.txt approach)